# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Solana War Room API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallets:
    get:
      operationId: listWallets
      tags: [wallets]
      summary: List all HD wallets
      responses:
        "200":
          description: List of wallets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Wallet"
    post:
      operationId: createWallets
      tags: [wallets]
      summary: Generate N new HD wallets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWalletsRequest"
      responses:
        "201":
          description: Created wallets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Wallet"
  /wallets/{walletId}:
    delete:
      operationId: deleteWallet
      tags: [wallets]
      summary: Delete a wallet and its accounts
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /wallets/export:
    post:
      operationId: exportWallets
      tags: [wallets]
      summary: Export all wallets with decrypted mnemonics and private keys
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportWalletsBody"
      responses:
        "200":
          description: Exported wallet data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExportedWallet"
        "401":
          description: Invalid password for one or more wallets
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    Wallet:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        encryptedMnemonic:
          type: string
        accountCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, encryptedMnemonic, accountCount, createdAt]
    CreateWalletsRequest:
      type: object
      properties:
        count:
          type: integer
          minimum: 1
          maximum: 50
        password:
          type: string
          description: Password used to encrypt the mnemonic on the server
      required: [count, password]
...
    ExportWalletsBody:
      type: object
      properties:
        password:
          type: string
          description: Decryption password for all wallets
      required: [password]
    ExportedWallet:
      type: object
      properties:
        walletId:
          type: integer
        walletName:
          type: string
        mnemonic:
          type: string
        accounts:
          type: array
          items:
            type: object
            properties:
              accountId:
                type: integer
              name:
                type: string
              publicKey:
                type: string
              privateKeyBase58:
                type: string
              hdPath:
                type: string
              hdIndex:
                type: integer
            required: [accountId, name, publicKey, privateKeyBase58, hdPath, hdIndex]
      required: [walletId, walletName, mnemonic, accounts]
...